load poll list inside navigation view and not in app
