+++ src/js/mixins/watchPolls.js
@@ -2,6 +2,7 @@
 import axios from '@nextcloud/axios'
 import { generateUrl } from '@nextcloud/router'
 import exception from '../Exceptions/Exceptions'
+import { emit } from '@nextcloud/event-bus'
 
 export const watchPolls = {
 	data() {
@@ -26,12 +27,8 @@ export const watchPolls = {
 			let dispatches = []
 			tables.forEach((item) => {
 				this.lastUpdated = Math.max(item.updated, this.lastUpdated)
-				// an updated poll table is reported
 				if (item.table === 'polls') {
-					if (this.$route.name !== 'publicVote') {
-						// load poll list only, when not in public poll
-						dispatches = [...dispatches, 'polls/list']
-					}
+					emit('update-polls')
 					if (item.pollId === parseInt(this.$route.params.id ?? this.$store.state.share.pollId)) {
 						// if current poll is affected, load current poll configuration
 						// load also options and votes
